requirements.txt
streamlit
pandas
pdfplumber
python-docx
spacy
nltk
sentence-transformers
scikit-learn
pymupdf
tqdm
